requirements.txt
langgraph>=0.0.15
langchain>=0.0.335
langchain-community>=0.0.13
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
streamlit>=1.28.0
python-dotenv>=1.0.0
firecrawl-py>=2.4.0
matplotlib>=3.7.1
pandas>=2.0.3
ollama>=0.1.5
PyPDF2>=3.0.1
requests>=2.31.0
beautifulsoup4>=4.12.2
numpy>=1.24.3
pydantic>=2.4.2